fix: 팔로우 및 좋아요 정보 샘플링 메서드 두 개를 1개로 통합 (#152)
