+++ src/main/java/com/devtraces/arterest/service/recommendation/initializing/RecommendationListInitializationService.java
@@ -1,0 +1,14 @@
+package com.devtraces.arterest.service.recommendation.initializing;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Service
+@Slf4j
+@RequiredArgsConstructor
+public class RecommendationListInitializationService {
+
+
+
+}
